Implement card click handling and matching logic
- Created onCardClicked function to handle the user's click on a card and prevents further clicks on the same card. or if two cards are already flipped by checking flippedCards.size and ensuring the clicked card is not already flipped.

If it's the first click and the timer hasn't started, the game timer is started by calling startTimer().

The clicked card is revealed by setting the image resource of the ImageView to the corresponding imageRes.

The flipped card along with its image resource (imageRes) is added to the flippedCards list for matching checks.

Once two cards are flipped, the checkForMatch() function is invoked to determine if the cards match.

implemented checkForMatch () fuction for pair matching.

- Created the checkForMatch function to compare the two flipped cards and check if they match.

If the two cards match the flippedCards list is cleared and the matchedPairs counter is incremented.

If all pairs are matched, the game completion message from strings.xml is displayed with the elapsed time, and the timer is stopped.

If the cards do not match, the cards are flipped back to their initial state after a delay of 1 second using handler.postDelayed().

The flippedCards list i is cleared after both matching and non-matching checks.

Implement game timer logic.

- Created startTimer function to begin the timer when the first card is flipped.

- The firstClick is set to false to prevent the timer from starting again with the click of the next flipped card.

- Used handler.postDelayed to repeatedly update the elapsedTime every second.

- The timer label (timerLabel.text) is updated with the formatted elapsed time.
The timer continues to run by posting every second until the game is reset or completed.